Add files via upload
I've attempted to implement the TODO item of fixing the min and max temperatures so that they no longer disappear on the first day of the forecast.  It seems to be working so far.

I've done this by using variables, TODAYS_MIN and TODAYS_MAX to store the min and max values of the "air_temp" from the BOM API as a default in the event that air_temperature_maximum or air_temperature_minimum contains 'n/a'.  As this changes over the day, the values are updated.  When the date changes, the TODAYS_MIN and TODAYS_MAX are reset.  LAST_DATE is used to track the change of date.